eventloop: Allow creating loop without loading "console" module
Allow creating an eventloop whose runtime does not have the "console"
module loaded into it by passing EnableConsole(false) into
NewEventLoop.

NewEventLoop now uses the functional options pattern and thus takes a
variadic number of option functions to modify its behavior and the
resulting event loop.  Currently, the only option that can be passed
to NewEventLoop is EnableConsole(bool), which allows controlling
whether NewEventLoop loads the "console" module into the runtime used
by the loop.  If no EnableConsole(false) option is passed to
NewEventLoop, the default is to load the "console" module.

Add two new tests, TestRunWithConsole and TestRunNoConsole, which
validate the presence or absence of the `console.log` command in the
runtime based on the use of EnableConsole(bool).