console: Allow configuring custom output for console
Allow configuring where `console` commands print to with the addition
of a new RequireWithPrinter(Printer) function which returns a
require.ModuleLoader that loads a `console` command into the given
runtime which writes to the supplied Printer.

Printer is a new interface type with methods, Log(string),
Warn(string) and Error(string).  A new private `printer` field has
been added to the Console type, and the appropriate Printer method is
called inside `console.log`, console.warn` and `console.error` to
print the final string after formatting its arguments.  A helper type
PrinterFunc can be used to create a Printer from a function with
signature `func(string)`.

If a custom Printer is not given, a default Printer is used which, for
all three methods, passes its string argument unmodified to log.Print,
thus preserving the existing behavior.

Added a new test TestConsoleWithPrinter which ensures that the
`console` command functions correctly when loaded into a runtime using
RequireWithPrinter and a custom Printer.