eventloop: Support adding timers & intervals to event loop from Go
Support adding timers & intervals to an event loop from Go using two
new methods SetTimeout and SetInterval, which return a Timer and
Interval, respectively.  Timers can be canceled by passing them to the
new ClearTimeout method or the existing clearTimeout JavaScript
function, and Intervals can be canceled by passing them to the new
ClearInterval method or the existing clearInterval JavaScript
function.  All four new methods SetTimeout, SetInterval, ClearTimeout
and ClearInterval can be called inside or outside of the event loop.

The Timer and Interval types are now exported to allow passing Timers
and Intervals between Go and JavaScript.

The event loop now stores jobs in a priority queue implemented using
the stdlib container/heap package.  The priority queue is sorted first
by earliest execution time and then by earliest added time.  All
access to the priority queue is protected with a mutex.  After
executing the last job, the event loop finds the "nearest" job in the
priority queue (i.e. the job that will execute the soonest) and sleeps
for the remaining time before waking up and executing the job.
Interval jobs are re-added to the priority queue by the event loop
with its new execution time just before executing its current
iteration.

Jobs that are added outside of the loop use a buffered "kick" channel
after the job has been added to the priority queue to wake up the
event loop, thus causing it to recalculate the nearest job and sleep
time.  Kicks sent via the "kick" channel are done in a separate
goroutine so as to not cause the caller to block.  The RunOnLoop,
SetTimeout and SetInterval methods use the "kick" channel as they may
be called outside of the run loop, whereas the setTimeout and
setInterval JavaScript functions do not as they are always executed
within the loop.  Since multiple kicks could be buffered before the
event loop wakes up, the event loop drains the "kick" channel after
receiving a kick to avoid spurious wake-ups causing the nearest job to
be recalculated needlessly.

Add new tests TestNativeTimeout, TestNativeClearTimeout,
TestNativeInterval, TestNativeClearInterval and TestClearInterval to
verify the new functionality.